trying to abstract out reusable code
