+++ example/generated-api-tree.ts
@@ -1,6 +1,6 @@
 import { IJimuApiOption } from "@jimengio/api-base";
-import { Id, EApiKind } from "./types";
-import { insertPublicHost } from "./configs";
+import { Id, EApiKind } from "../src/types";
+import { insertPublicHost } from "../src/configs";
 import {
   ajaxGet,
   ajaxDelete,
@@ -14,7 +14,7 @@ import {
   dynamicPost,
   dynamicPut,
   dynamicDelete,
-} from "./funcs";
+} from "../src/funcs";
 
 /** Generated by generator */
