+++ src/generator/interface-collector.ts
