+++ src/index.ts
@@ -1,3 +1,1 @@
 export { configureSeedHosts } from "./configs";
-
-export * from "./generated-api-tree";
